feat(Plugin): Add plugin
ProjectMgt: plugin for (un)assign projects from/to items
