+++ MiAZ/data/docs/VERSION
@@ -1,1 +1,1 @@
-0.0.4+build.285
+0.0.4+build.304
