+++ MiAZ/data/plugins/projmgt.plugin
@@ -1,0 +1,10 @@
+[Plugin]
+Module=projmgt
+Name=MiAZProjectMgt
+Loader=Python3
+Description=(Un)assign projects from/to items
+Authors=Tomás Vírseda <tomasvirseda@gmail.com>
+Copyright=Copyright © 2023 Tomás Vírseda
+Website=http://github.com/t00m/MiAZ
+Help=http://github.com/t00m/MiAZ/README.adoc
+Version=0.1
