+++ MiAZ/frontend/desktop/actions.py
@@ -467,101 +467,6 @@ class MiAZActions(GObject.GObject):
         dialog.show()
 
 
-    def project_assign(self, item_type, items):
-        def dialog_response(dialog, response, dropdown, items):
-            if response == Gtk.ResponseType.ACCEPT:
-                self.projects = self.backend.projects
-                pid = dropdown.get_selected_item().id
-                docs = []
-                for item in items:
-                    docs.append(os.path.basename(item.id))
-                self.projects.add_batch(pid, docs)
-                workspace = self.app.get_widget('workspace')
-                workspace.update()
-            dialog.destroy()
 
-        self.factory = self.app.get_factory()
-        self.config = self.backend.conf
-        i_type = item_type.__gtype_name__
-        box = self.factory.create_box_vertical(spacing=6, vexpand=True, hexpand=True)
-        dropdown = self.factory.create_dropdown_generic(Project)
-        self.config[i_type].connect('used-updated', self.dropdown_populate, dropdown, item_type, False, False)
-        self.dropdown_populate(self.config[i_type], dropdown, Project, any_value=False)
-        btnManage = self.factory.create_button('miaz-res-manage', '')
-        btnManage.connect('clicked', self.manage_resource, Configview['Project'](self.app))
-        label = self.factory.create_label('Assign the following documents to this project: ')
-        frame = Gtk.Frame()
-        cv = MiAZColumnViewMassProject(self.app)
-        cv.get_style_context().add_class(class_name='caption')
-        cv.set_hexpand(True)
-        cv.set_vexpand(True)
-        citems = []
-        self.projects = self.backend.projects
-        for item in items:
-            projects = ', '.join([self.projects.description(pid) for pid in self.projects.assigned_to(item.id)])
-            citems.append(File(id=item.id, title="<b>%s</b>" % projects))
-        cv.update(citems)
-        frame.set_child(cv)
-        hbox = self.factory.create_box_horizontal(hexpand=False, vexpand=False)
-        hbox.append(label)
-        hbox.append(dropdown)
-        hbox.append(btnManage)
-        box.append(hbox)
-        box.append(frame)
-        dialog = self.factory.create_dialog_question(self.app.win, 'Assign to a project', box, width=1024, height=600)
-        dialog.connect('response', dialog_response, dropdown, items)
-        dialog.show()
-
-    def project_withdraw(self, item_type, items):
-        def dialog_response(dialog, response, dropdown, items):
-            if response == Gtk.ResponseType.ACCEPT:
-                self.projects = self.backend.projects
-                pid = dropdown.get_selected_item().id
-                docs = []
-                for item in items:
-                    docs.append(os.path.basename(item.id))
-                self.projects.remove_batch(pid, docs)
-                workspace = self.app.get_widget('workspace')
-                workspace.update()
-            dialog.destroy()
-
-        self.factory = self.app.get_factory()
-        self.config = self.backend.conf
-        i_type = item_type.__gtype_name__
-        box = self.factory.create_box_vertical(spacing=6, vexpand=True, hexpand=True)
-        dropdown = self.factory.create_dropdown_generic(Project)
-        self.config[i_type].connect('used-updated', self.dropdown_populate, dropdown, item_type, False, False)
-
-        # Get projects
-        projects = set()
-        for item in items:
-            for project in self.backend.projects.assigned_to(item.id):
-                projects.add(project)
-
-        self.dropdown_populate(self.config[i_type], dropdown, Project, any_value=False, only_include=list(projects))
-        btnManage = self.factory.create_button('miaz-res-manage', '')
-        btnManage.connect('clicked', self.manage_resource, Configview['Project'](self.app))
-        label = self.factory.create_label('Withdraw the following documents from this project: ')
-        frame = Gtk.Frame()
-        cv = MiAZColumnViewMassProject(self.app)
-        cv.get_style_context().add_class(class_name='caption')
-        cv.set_hexpand(True)
-        cv.set_vexpand(True)
-        citems = []
-        self.projects = self.backend.projects
-        for item in items:
-            projects = ', '.join([self.projects.description(pid) for pid in self.projects.assigned_to(item.id)])
-            citems.append(File(id=item.id, title="<b>%s</b>" % projects))
-        cv.update(citems)
-        frame.set_child(cv)
-        hbox = self.factory.create_box_horizontal(hexpand=False, vexpand=False)
-        hbox.append(label)
-        hbox.append(dropdown)
-        hbox.append(btnManage)
-        box.append(hbox)
-        box.append(frame)
-        dialog = self.factory.create_dialog_question(self.app.win, 'Assign to a project', box, width=1024, height=600)
-        dialog.connect('response', dialog_response, dropdown, items)
-        dialog.show()
